feat: add store data, fixed index grid for Desktop and added Mobile draft
