+++ store/home.js
@@ -1,0 +1,23 @@
+import universeSimulation from "../public/images/categories/universeSimulation.jpg";
+import electricity from "../public/images/categories/electricity.jpg";
+import glassBox from "../public/images/categories/glassBox.jpg";
+import temporalDrift from "../public/images/categories/temporalDrift.jpg";
+import theThinBlueLie from "../public/images/categories/theThinBlueLie.jpg";
+
+const categories = [
+  {
+    name: "moving through space and time",
+    img: universeSimulation,
+    link: "/moving-through-space-and-time",
+  },
+  { name: "Electricity", img: electricity, link: "/electricity" },
+  { name: "Glass Box", img: glassBox, link: "/glass-box" },
+  { name: "Temporal Drift", img: temporalDrift, link: "/temporal-drift" },
+  {
+    name: "The Thin Blue Lie",
+    img: theThinBlueLie,
+    link: "/police-brutality",
+  },
+];
+
+export { categories };
